# perf: Allocation reduction