@emersion 2023-12-22
In principle these changes look like a good idea to me! I haven't looked at the code yet, will get to it in the next few days.