@56KBs 2023-12-23
Here are a couple examples of the changes making use of `strings.Cut` from Go 1.18+.

`authres/parse.go`
```go
func parseParam(s string) (k string, v string, err error) {
	key, value, ok := strings.Cut(s, "=")
	if !ok {
		return "", "", errors.New("msgauth: malformed authentication method and value")
	}

	return strings.TrimSpace(strings.ToLower(key)), strings.TrimSpace(value), nil
}
```

`dkim/verify.go`
```go
	// Parse algos
	keyAlgo, hashAlgo, ok := strings.Cut(stripWhitespace(params["a"]), "-")
	if !ok {
		return verif, permFailError("malformed algorithm name")
	}
```