@emersion 2023-12-26
> These changes would look even more simple through the use of strings.Cut, however this was added in Go 1.18 and I didn't want to force a minimum version bump for the sake of this change (Although 1.18 is now out of support) - If this would be preferred I can do this instead.

I would be completely okay with bumping our minimum required Go version to 1.18.